actually run the commands
